pyproject.toml
[build-system]
requires = [
    "setuptools",
    "Cython",
    "numpy"
]
build-backend = "setuptools.build_meta"

[project]
name = "happy-clouds"
version = "2025.1.0"
authors = [
    { name = "Larry Page", email = "lawrence.page@live.com" },
]
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.13"
license = { text = "MIT License" }
dependencies = [ 
    "pillow",
    "numpy",
    "matplotlib",
    "webcolors"
]

[project.urls]
Homepage = "https://github.com/lpage30/happy-clouds"

[project.scripts]
generate_imagecloud = "clis.cli_functions:generate_imagecloud"
layout_imagecloud = "clis.cli_functions:layout_imagecloud"
generate_textcloud = "clis.cli_functions:generate_textcloud"
layout_textcloud = "clis.cli_functions:layout_textcloud"
generate_textimagecloud = "clis.cli_functions:generate_textimagecloud"
layout_textimagecloud = "clis.cli_functions:layout_textimagecloud"
generate_mixeditemcloud = "clis.cli_functions:generate_mixeditemcloud"
layout_mixeditemcloud = "clis.cli_functions:layout_mixeditemcloud"
show_all_text_fonts = "clis.show_all_fonts:show_all_fonts"
cloud_usage_md = "clis.generate_md_help:generate_md_help"
generate_weighted_text = "clis.generate_random_weighted_text_csv:main"

[tool.setuptools.packages.find]
where = ["src"]